feat: speed: Measure vehicle speed using pin-change IRQ on VSS
This commit implements vehicle speed measurement using a pin-change
interrupt on the vehicle speed sensor (VSS) line connected to pin D7

The VSS emits ~4250 pulses per mile[^1] under normal conditions. As
with NEP, attach a pin-change interrupt; if when the interrupt
activates, the pin is high, recalculate the vehicle speed by checking
the time since the last revolution.

Also, reflect the new measurement details in
`debugUSB_printOEMsignals` (the OEM signal echoing mode).

Notes:

- A more precise value can be derived from the Honda DTC manual [^2]:

  > The vehicle speed sensor (VSS) is attached to the transmission
  > housing. The VSS outputs a pulsing signal for every revolu- tion
  > of the differential gear via the speedometer gear.

[^1]:
https://www.insightcentral.net/threads/vss-vehicle-speed-sensor-output.13597/post-191883
describes it initially as 4550 pulses, but some experimentation and
comparison to a stock speedo led to the 4250 number now used.

[^2]: https://downloads.laboratoryb.org/insight/other_manuals/HONDA_DTC.pdf#page=106

Signed-off-by: Martin Kennedy <hurricos@gmail.com>