[WIP] feat: mode1: Blend in a new map for IMA power based on TPS+VSS+BRAKE
With this commit, blend a mapped "suggested" CMDPWR value based on the
vehicle speed and TPS into the actual CMDPWR in LiControl mode 1.

The mapping is constructed so as to promote the use of IMA power at
low throttle. This reduces the fatigue of using the IMA joystick,
essentially allows regen to occur (for driving comfort) as soon as
driving begins, and (for the CVT specifically) generates a driving
habit which links IMA torque to engine torque, sparing the CVT from
miscalculating total torque.

Work-in-progress, as it probably needs to be gated behind
TRANSMISSION_IS_CVT.

Also, I need to find out why brake detection isn't working consistently.